safeloader: split into smaller modules
This splits the somewhat large (and more complex than necessary)
safeloader.py file into a number of smaller, more "self contained"
modules.
